change vote type to float

+++ app/src/main/java/com/limamauricio/popularmovies/model/Movie.java
@@ -32,7 +32,7 @@ public class Movie implements Serializable {
     @Getter
     @Setter
     @SerializedName("vote_average")
-    private double voteAverage;
+    private float voteAverage;
 
     @Getter
     @Setter
